Changed the formal charge from Double to Integer
git-svn-id: https://cdk.svn.sourceforge.net/svnroot/cdk/cdk/branches/cdk-1.2.x@14050 eb4e18e3-b210-0410-a6ab-dec725e4b171

+++ src/main/org/openscience/cdk/debug/DebugAdductFormula.java
@@ -46,8 +46,8 @@ public class DebugAdductFormula extends AdductFormula implements IAdductFormula 
 	    return super.contains(isotope);
     }
 
-	public Double getCharge() {
-		Double charge = super.getCharge();
+	public Integer getCharge() {
+		Integer charge = super.getCharge();
 		logger.debug("Getting charge: ", charge);
 	    return charge;
     }
@@ -67,7 +67,7 @@ public class DebugAdductFormula extends AdductFormula implements IAdductFormula 
 	    return super.isotopes();
     }
 
-	public void setCharge(Double charge) {
+	public void setCharge(Integer charge) {
 	    logger.debug("Setting the charge to: ", charge);
 	    super.setCharge(charge);
     }
